Refactor app_is_not_removed helper
