+++ saleor/graphql/app/mutations/app_token_create.py
@@ -11,7 +11,7 @@ from ...core.mutations import ModelMutation
 from ...core.types import AppError, BaseInputObjectType
 from ...utils import get_user_or_app_from_context, requestor_is_superuser
 from ..types import AppToken
-from .utils import app_is_not_removed
+from ..utils import app_is_not_removed
 
 
 class AppTokenInput(BaseInputObjectType):
